% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_content.R
\name{create_page}
\alias{create_page}
\title{wikimedia api page creation (single pages)
 
helper function to do the actual api requests for page 
and category-page creation}
\usage{
create_page(url, p_title, p_text, category, token)
}
\arguments{
\item{url}{a URL body}

\item{p_title}{page title string of new page}

\item{p_text}{page content string of new page}

\item{category}{switch to decide, if the page should be
created as category-page}

\item{token}{action token to perform the request}
}
\value{
TRUE
}
\description{
wikimedia api page creation (single pages)
 
helper function to do the actual api requests for page 
and category-page creation
}
