% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_content.R
\name{create_pages}
\alias{create_pages}
\title{wikimedia api page creation
 
Create pages or category-pages on a wikimedia instance.}
\usage{
create_pages(url, p_title, p_text, category = FALSE)
}
\arguments{
\item{url}{a URL body}

\item{p_title}{vector with page title strings of new pages}

\item{p_text}{vector with page content strings of new pages}

\item{category}{switch to decide, if the pages should be
created as category-pages}
}
\value{
TRUE
}
\description{
wikimedia api page creation
 
Create pages or category-pages on a wikimedia instance.
}
