% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_content.R
\name{get_action_token}
\alias{get_action_token}
\title{request token for api action as signed in user}
\usage{
get_action_token(url)
}
\arguments{
\item{url}{a URL body}
}
\value{
a token string
}
\description{
helper function to request a user action token
}
