% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{get_prelogin_token}
\alias{get_prelogin_token}
\title{request token to start client login}
\usage{
get_prelogin_token(url, user)
}
\arguments{
\item{url}{a URL body}

\item{user}{a username of a registered user}
}
\value{
a token string
}
\description{
helper function to request a user login token
}
