% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login}
\alias{login}
\title{wikimedia api user login}
\usage{
login(url, user, pw)
}
\arguments{
\item{url}{a URL body}

\item{user}{a username of a registered user}

\item{user}{the password of said user}
}
\value{
TRUE
}
\description{
Login to a wikimedia instance to trigger api requests
as a registered user. This function only allows the 
very basic login with username and password. Wikimedia
setups that require more sophisticated login methods 
are not supported.
}
