% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/content.R
\name{page_content}
\alias{page_content}
\title{Retrieves MediaWiki page content}
\usage{
page_content(language = NULL, project = NULL, domain = NULL, page,
  properties = c("text", "revid"))
}
\arguments{
\item{language}{The language code of the project you wish to query,
if appropriate.}

\item{project}{The project you wish to query ("wikiquote"), if appropriate.
Should be provided in conjunction with \code{language}.}

\item{domain}{as an alternative to a \code{language} and \code{project} combination,
you can also provide a domain ("rationalwiki.org") to the URL constructor, allowing
for the querying of non-Wikimedia MediaWiki instances.}

\item{page}{The title of the page you want to retrieve}

\item{properties}{Properties associated with the page, namely "text" (the actual, HTML content)
and "revid" (the revision ID of the current version), the latter of which allows you to use
\code{\link{wiki_revision}} to retrieve the underlying wikitext.}
}
\description{
wiki_page retrieves the DOM of a particular MediaWiki page,
as a HTML blob inside a JSON object.
}
\seealso{
\code{\link{wiki_diff}} for retrieving 'diffs' between revisions,
\code{\link{wiki_revision}} for retrieving the text of specified revisions.
}

