% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_response}
\alias{parse_response}
\title{parse_response: Parse WikipediR responses internally}
\usage{
parse_response(x)
}
\arguments{
\item{x}{result from a WikipediR query}
}
\description{
Response parser
}
\details{
Should not be externally used
}
