% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wiki_call}
\alias{wiki_call}
\title{Basic calling function}
\usage{
wiki_call(URL, ...)
}
\arguments{
\item{URL}{the URL to pass to GET. Set by the calling function.}

\item{...}{any other parameters to pass through to GET.}
}
\description{
Function used as a component of other package functions that handles the actual
connection to the MediaWiki API
}

