% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{wiki_conClass-class}
\alias{wiki_conClass}
\alias{wiki_conClass-class}
\title{a reference class for connector objects}
\description{
Class for connector objects. Currently (relatively) a placeholder; will need to be fleshed out
more once oAuth authentication and logins are live.
}
\section{Fields}{

\describe{
\item{\code{URL}}{the base URL of the connector object. A length-one character vector.}

\item{\code{CurlOpts}}{options to pass through to RCurl; a list.}
}}

