\name{wiki_page}
\alias{wiki_page}
\title{
Retrieves MediaWiki page content
}
\description{
wiki_page retrieves the DOM of a particular MediaWiki page, as a HTML blob inside a JSON object.
}
\usage{
wiki_page(con, page, properties = c("text","revid"))

}
\arguments{
  \item{con}{
  A connector object, generated by \code{\link{wiki_con}}, that corresponds to the project you're trying to query.
}
  \item{page}{
The title of the page you want to retrieve
}
  \item{properties}{
Properties associated with the page, namely "text" (the actual, HTML content) and "revid" (the revision ID of the current version), the latter of which allows you to use \code{\link{wiki_revision}} to retrieve the underlyinwikitext.
}
}

\seealso{
\code{\link{wiki_diff}} for retrieving 'diffs' between revisions, \code{\link{wiki_revision}} for retrieving the text of specified revisions.
}