% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WikipediR.R
\docType{package}
\name{WikipediR}
\alias{WikipediR}
\alias{WikipediR-package}
\title{A client library for MediaWiki's API}
\description{
This package provides functions for accessing the MediaWiki API, either for
Wikimedia projects or any other MediaWiki instance. For more information, see the
\href{https://cran.r-project.org/web/packages/WikipediR/vignettes/WikipediR.html}{vignette}.
}
\seealso{
The \href{https://CRAN.R-project.org/package=WikipediR/vignettes/WikipediR.html}{package vignette}.
}

