\name{wikis}
\alias{wikis}
\docType{data}
\title{
Active Wikimedia production wikis
}
\description{
A data frame of active Wikimedia production wikis - in other words, all acceptable permutations of language and project code.
}
\usage{data(wikis)}
\format{
  A data frame with 688 observations on the following 2 variables.
  \describe{
    \item{\code{code}}{a character vector consisting of language codes (e.g. "en" for the English-language projects)}
    \item{\code{project}}{a character vector consisting of project types (e.g. "wikipedia")}
  }
}
