\name{wiki_diff}
\alias{wiki_diff}
\title{
Generates a "diff" between a pair of revisions
}
\description{
wiki_diff generates a diff between two revisions in a MediaWiki page. This is provided as an XML-parsable blob inside the returned JSON object.}
\usage{
  wiki_diff(con, revID, properties = c("ids", "flags", "timestamp",
                                       "user", "userid", "size",
                                       "sha1", "contentmodel", "comment",
                                       "parsedcomment", "tags", "flagged"),
            direction = c("prev", "next", "cur"))
}
\arguments{
  \item{con}{
  A connector object, generated by \code{\link{wiki_con}}, that corresponds to the project you're trying to query.

}
  \item{revID}{
  The revision ID of the "start" revision
}
  \item{properties}{
  Properties you're trying to retrieve about that revision, should you want to; options include "ids" (the revision ID of the revision...which is pointless), "flags" (whether the edit was 'minor' or not), "timestamp" (the timestamp of the revision, which can be parsed with \code{\link{wiki_timestamp}}), "user" (the username of the person who made that edit), "userid" (the userID of the person who made the edit), "size" (the size, in uncompressed bytes, of the revision), "sha1" (the SHA-1 hash of the revision text), "contentmodel" (the content model of the page, usually "wikitext"), "comment" (the edit summary associated with the revision), "parsedcomment" (the same, but parsed, generating HTML from any wikitext in that comment), "tags" (any tags associated with the edit) and "flagged" (the edit's status under Flagged Revisions)}
  \item{direction}{
  The direction you want the diff to go in from the revisionID you have provided. Options are "prev" (compare to the previous revision on that page), "next" (compare to the next revision on that page) and "cur" (compare to the current, extant version of the page).
}
}
\seealso{

  \code{\link{wiki_con}}, \code{\link{wiki_timestamp}}
}
\examples{
  #Create connector
  ih_connector <- wiki_con(language = "en", project = "wikipedia")
  
  #Ask for a diff between revision 602943163 and the next-dated revision on that page, also retrieving the username of the person who /made/ said revision
  ih_pagdiff <- wiki_diff(con = ih_con, revID = 602943163, properties = "user", direction = "next")
}