% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wiki_timestamp}
\alias{wiki_timestamp}
\title{Timestamp conversion function}
\usage{
wiki_timestamp(timestamp, to_wiki = FALSE)
}
\arguments{
\item{timestamp}{The timestamp(s) you wish to convert, provided as a vector. If the intent
is to convert them into MediaWiki timestamps, they must be POSIX-compliant.}

\item{to_wiki}{Whether the intent is to convert them from POSIX to MediaWiki. Set to
FALSE by default}
}
\description{
\code{\link{wiki_timestamp}} converts MediaWiki timestamps to POSIX-compliant
timestamps, and vice-versa.
}

