% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wikis}
\alias{wikis}
\title{Valid Wikimedia projects}
\format{A data frame with 688 rows and 2 variables}
\description{
A data.frame of valid language code and project type permutations in Wikimedia's
MediaWiki instances. Columns are:

\itemize{
\item code. The language code (en, ar, de...)
\item project. The project type (wikipedia, wiktionary...)
}
}

